+++ scripts/install.sh
@@ -1,8 +1,8 @@
 #!/bin/bash
 
-if [ -x "$(which python)" ]; then
-    rm -r py2
+./scripts/clean.sh
 
+if [ -x "$(which python)" ]; then
     virtualenv --clear --no-site-packages -p python py2
 
     ./py2/bin/pip install coverage
@@ -16,8 +16,6 @@ if [ -x "$(which python)" ]; then
     ./py2/bin/pip install -e .[test]
 fi
 if [ -x "$(which python3)" ]; then
-    rm -r py3
-
     virtualenv --clear --no-site-packages -p python3 py3
 
     ./py3/bin/pip install coverage
@@ -29,19 +27,4 @@ if [ -x "$(which python3)" ]; then
     ./py3/bin/pip install https://github.com/conestack/plumber/archive/master.zip
     ./py3/bin/pip install https://github.com/conestack/node/archive/master.zip
     ./py3/bin/pip install -e .[test]
-fi
-if [ -x "$(which pypy)" ]; then
-    rm -r pypy
-
-    virtualenv --clear --no-site-packages -p pypy pypy
-
-    ./pypy/bin/pip install coverage
-    ./pypy/bin/pip install zope.interface
-    ./pypy/bin/pip install zope.lifecycleevent
-    ./pypy/bin/pip install zope.component
-    ./pypy/bin/pip install zope.deprecation
-    ./pypy/bin/pip install https://github.com/conestack/odict/archive/master.zip
-    ./pypy/bin/pip install https://github.com/conestack/plumber/archive/master.zip
-    ./pypy/bin/pip install https://github.com/conestack/node/archive/master.zip
-    ./pypy/bin/pip install -e .[test]
 fi